refactor[blog-list]: bring api call out of the component
